Add validations on endpoints

+++ src/main/java/com/booking/project/service/interfaces/IReservationService.java
@@ -27,7 +27,7 @@ public interface IReservationService {
 
     List<ReservationDTO> filterGuestReservations(String title, LocalDate startDate, LocalDate endDate, ReservationStatus reservationStatus, Long guestUserId);
 
-    List<ReservationDTO> filterHostReservations(String title, LocalDate startDate, LocalDate endDate, ReservationStatus reservationStatus, Integer hostUserId);
+    List<ReservationDTO> filterHostReservations(String title, LocalDate startDate, LocalDate endDate, ReservationStatus reservationStatus, Long hostUserId);
 
     Reservation updateStatus(Long id, ReservationStatus reservationStatus) throws Exception;
